requirements.txt
asgiref==3.5.2
Django==4.0.6
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-modeltranslation==0.18.2
gunicorn==20.1.0
Pillow==9.2.0
# psycopg2-binary==2.9.3
python-decouple==3.6
six==1.16.0
sqlparse==0.4.2
transliterate==1.10.2
